+++ test/nunit/TestClass2.cs
@@ -2,7 +2,7 @@ using System;
 using NUnit.Framework;
 using Shouldly;
 
-namespace XunitTests
+namespace NunitTests
 {
     public class TestClass2
     {