+++ test/xunit/TestClass1.cs
@@ -4,7 +4,6 @@ using Xunit;
 
 namespace XunitTests
 {
-
     public class TestClass1
     {
         [Fact]
@@ -22,7 +21,7 @@ namespace XunitTests
         [Fact]
         public void Fail()
         {
-            (1+1).ShouldBe(2);
+            (1+1).ShouldBe(3);
         }        
     }
 }